Removing upgrade folder from robottelo tests
